requirements.txt
Django==4.2.7
django-environ==0.11.2
django-ninja==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==12.0.0
python-dotenv==1.0.0